Update echo-server.ts
res.channel_data is already an object, so trying to parse it was causing an issue. Also I don't think we want to send member.socketId down to the client since it reveals the socket ID down to other people in the channel. Not sure if that matters security wise but just seemed unneeded.