Fix checking for http via host option.
When checking for which http protocol to use, the "host" option in the json file is tested to see if it starts with `https` or `http`

However, if the host option DOES start with either of those, the script throws a `Error: getaddrinfo ENOTFOUND` error. See [tech info here](http://stackoverflow.com/a/17691556/659786).

So removing the protocol from the `host` options means we must use something else to check for http(s). This patch uses the `protocol` option in the json file. First suggested by https://github.com/Thaars