Allow wildcard and multi domain authHost option
As described in #185 as well as #135:
- allow for wildcard domain declaration in authHost option (e.g. authHost = '.myapp.com') so one can use the laravel-echo-server with a multi tenant/subdomain system 
- as well as a multi domain system, where different authentication servers are used. 
The currently appropriate authentication domain is retrived from the referer and matched with the (optional: wildcard) domain declaration from the authHost option as a whitelist (which can be either an array or a string). In case none is matched with the referer, the last one in the array is taken (in case of string, the only one). 
@tlaverdure @barryvdh please check. I have done my best to write this up in TS, but I don't have a compile system for this setup, so I have only tested it in vanilla JS and then afterwards added the TS specific types for this PR.
Thanks!