If cookie are sent within the message, use them.
Since cookies are always retrieved from the first request payload,
it's really hard to handle authentication on the client side.

One example, if the socket connexion is opened before the user is logged,
the socket will never be able to know about the logged state.

By using the options.headers['Cookie'] value as default (if set) we let
the user bypass cookies and handle the full state from the client side.

Signed-off-by: Stéphane HULARD <s.hulard@chstudio.fr>